Use qreal instead of int to store tablet pad ring positions
While looking through the libinput API, we noticed that the
libinput_event_tablet_pad_get_ring_position actually returns a double -
not an integer. My own hardware doesn't emit anything this low, but it's
possible something does.

More importantly, the Wayland protocol accepts a double for ring
position events. So if there is hardware that allows shifting a ring in
quarter or half degrees, applications should be able to handle those
now.

+++ autotests/integration/kwin_wayland_test.h
@@ -1079,7 +1079,7 @@ void touchUp(qint32 id, quint32 time);
 void tabletPadButtonPressed(quint32 button, quint32 time);
 void tabletPadButtonReleased(quint32 button, quint32 time);
 void tabletPadDialEvent(double delta, int number, quint32 time);
-void tabletPadRingEvent(int position, int number, quint32 group, quint32 mode, quint32 time);
+void tabletPadRingEvent(qreal position, int number, quint32 group, quint32 mode, quint32 time);
 void tabletToolButtonPressed(quint32 button, quint32 time);
 void tabletToolButtonReleased(quint32 button, quint32 time);
 void tabletToolProximityEvent(const QPointF &pos, qreal xTilt, qreal yTilt, qreal rotation, qreal distance, bool tipNear, qreal sliderPosition, quint32 time);
